feat(pkg): renamed java package
