+++ backend/java/build.gradle
@@ -5,7 +5,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.4'
 }
 
-group = 'com.michaelyi'
+group = 'com.michaelyi.personalwebsite'
 version = '2.0.0'
 
 java {
@@ -37,7 +37,7 @@ dependencies {
 
 jar {
     manifest {
-        attributes 'Main-Class': 'com.michaelyi.App'
+        attributes 'Main-Class': 'com.michaelyi.personalwebsite.App'
     }
 }
