--- backend/java/src/main/java/com/michaelyi/App.java
@@ -1,11 +1,0 @@
-package com.michaelyi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class App {
-    public static void main(String[] args) {
-        SpringApplication.run(App.class, args);
-    }
-}
