--- backend/java/src/main/java/com/michaelyi/auth/LoginRequest.java
@@ -1,6 +1,0 @@
-package com.michaelyi.auth;
-
-public record LoginRequest(
-        String password
-) {
-}
