--- backend/java/src/main/java/com/michaelyi/auth/UnauthorizedException.java
@@ -1,7 +1,0 @@
-package com.michaelyi.auth;
-
-public class UnauthorizedException extends RuntimeException {
-    public UnauthorizedException() {
-        super("Unauthorized");
-    }
-}
