--- backend/java/src/main/java/com/michaelyi/auth/User.java
@@ -1,63 +1,0 @@
-package com.michaelyi.auth;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-import static com.michaelyi.util.Constants.ADMIN_EMAIL;
-import static com.michaelyi.util.Constants.ADMIN_ROLE;
-
-public class User implements UserDetails {
-    private String password;
-
-    public User(String password) {
-        this.password = password;
-    }
-
-    public User() {
-    }
-
-    @Override
-    @JsonIgnore
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        SimpleGrantedAuthority role = new SimpleGrantedAuthority(
-                String.format("ROLE_%s", ADMIN_ROLE));
-
-        return List.of(role);
-    }
-
-    @Override
-    @JsonIgnore
-    public String getPassword() {
-        return password;
-    }
-
-    @Override
-    public String getUsername() {
-        return ADMIN_EMAIL;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
