--- backend/java/src/main/java/com/michaelyi/auth/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package com.michaelyi.auth;
-
-public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException() {
-        super("User not found.");
-    }
-}
