+++ backend/java/src/main/java/com/michaelyi/personalwebsite/App.java
@@ -1,0 +1,11 @@
+package com.michaelyi.personalwebsite;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class App {
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
+}
