+++ backend/java/src/main/java/com/michaelyi/personalwebsite/auth/LoginRequest.java
@@ -1,0 +1,6 @@
+package com.michaelyi.personalwebsite.auth;
+
+public record LoginRequest(
+        String password
+) {
+}
