+++ backend/java/src/main/java/com/michaelyi/personalwebsite/auth/UnauthorizedException.java
@@ -1,0 +1,7 @@
+package com.michaelyi.personalwebsite.auth;
+
+public class UnauthorizedException extends RuntimeException {
+    public UnauthorizedException() {
+        super("Unauthorized");
+    }
+}
