+++ backend/java/src/main/java/com/michaelyi/personalwebsite/auth/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package com.michaelyi.personalwebsite.auth;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException() {
+        super("User not found.");
+    }
+}
