--- backend/java/src/main/java/com/michaelyi/post/PostExceptionHandler.java
@@ -1,36 +1,0 @@
-package com.michaelyi.post;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-import java.util.NoSuchElementException;
-
-@ControllerAdvice
-public class PostExceptionHandler {
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<String> handleException(
-            IllegalArgumentException e
-    ) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(JsonProcessingException.class)
-    public ResponseEntity<String> handleException(
-            JsonProcessingException e
-    ) {
-        return new ResponseEntity<>(
-                e.getMessage(),
-                HttpStatus.INTERNAL_SERVER_ERROR
-        );
-    }
-
-    @ExceptionHandler(NoSuchElementException.class)
-    public ResponseEntity<String> handleException(
-            NoSuchElementException e
-    ) {
-        return new ResponseEntity<>(e.getMessage(), HttpStatus.NOT_FOUND);
-    }
-}
